Fix RpcHistory output

+++ src/GUI/Patches/LobbyTimerPatch.cs
@@ -49,8 +49,8 @@ public class TimerTextStringPatch
     {
         if (__instance.name != "WarningText") return true;
         int timer = __instance.GetSecondsRemaining();
-        int minutes = (int)timer / 60;
-        int seconds = (int)timer % 60;
+        int minutes = timer / 60;
+        int seconds = timer % 60;
         string suffix = $" {minutes:00}:{seconds:00}";
         if (timer <= 60) suffix = Color.red.Colorize(suffix);
         __instance.text.text = string.Format(__instance.format, suffix);
